pyproject.toml
[tool.poetry]
name = "playwright_ui_db_framework_poetry"
version = "0.1.0"
description = "Playwright-based test framework with UI and DB validation"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
playwright = "*"
pytest = "*"
pytest-playwright = "*"
pytest-html = "*"
pytest-xdist = "*"
psycopg2-binary = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
